Added ability to check if the server is in a shutdown process Updated sending of disconnect message after the channel has been removed
The issue we encountered in the nabu development stack:

The server was updated due to deployment and needed a reload.
Reloading entails stopping the current server.
Stopping the server kills all the connections.
Killing the connections triggers the websocket disconnect listener to do some final actions
Doing the final actions triggered a "get" on the repository
The get triggered the repository entry to see "hey the server has been updated, let me reload that for you"
The reload entails stopping the current server.
...
ad infinitum

The root cause is that the connection was not removed from the list of connections _before_ the event was published
If the event is handled synchronously and the connection is still there, it goes on forever
Simply removing the connection from the list before triggering the event should solve it